requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.6.0
attrs==23.2.0
bitarray==3.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
crayons==0.4.0
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
cytoolz==1.0.1
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
filelock==3.17.0
Flask==3.0.2
flask-cors==5.0.1
flutter==0.1
fonttools==4.51.0
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn
h11==0.14.0
hexbytes==1.3.0
huggingface-hub==0.29.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mongoose==0.0.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
Naked==0.1.32
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parsimonious==0.10.0
pefile==2023.2.7
piexif==1.1.3
pillow==10.2.0
ply==3.11
pocketsphinx==5.0.3
preshed==3.0.9
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pynput==1.8.1
pyparsing==3.1.2
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
pyunormalize==16.0.0
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rlp==4.1.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.19.0
sentence-transformers==3.4.1
setuptools==69.1.1
shellescape==3.8.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.4.8
stegano==0.11.5
sympy==1.13.1
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.2
transformers==4.49.0
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
types-requests==2.32.0.20250306
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
web3==7.8.0
webdriver-manager==4.0.1
websockets==13.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.18.3
youtube-transcript-api==0.6.2
